split up the types.ts into rating and messages
